+++ rolabesti/models/id3.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class ID3Tags(str, Enum):
+    artist = "artist"
+    title = "title"
+    album = "album"
+    genre = "genre"
