+++ rolabesti/models/sortings.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class Sortings(str, Enum):
+    asc = "asc"
+    desc = "desc"
+    random = "random"
